Refactor check to support different versions easily
