+++ kafka_consumer/tests/test_kafka_consumer.py
@@ -23,7 +23,7 @@ def test_check_kafka(aggregator, kafka_instance):
     """
     Testing Kafka_consumer check.
     """
-    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, {})
+    kafka_consumer_check = KafkaCheck('kafka_consumer', {}, [{}])
     kafka_consumer_check.check(kafka_instance)
 
     for name, consumer_group in iteritems(kafka_instance['consumer_groups']):
